pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "capellini"
version = "0.1.0"
description = "CRISPR-Abundance Phage-Evidence Linkage for Leveraging Interaction Network Inference"
requires-python = ">=3.10"
dependencies = [
    "biopython>=1.81",
    "numpy>=1.24",
    "pandas>=2.0",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "matplotlib>=3.7",
    "seaborn>=0.12",
    "networkx>=3.1",
    "tqdm>=4.65",
    "rich>=13.0",
    "pyyaml>=6.0",
    "questionary>=2.0",
]

[project.scripts]
capellini = "capellini.cli:main"

[project.optional-dependencies]
dev = ["pytest", "black", "ruff"]
docs = [
    "sphinx>=7.0",
    "sphinx-rtd-theme>=2.0",
    "myst-parser>=2.0",
    "sphinx-autobuild>=2024.0",
]

[tool.setuptools.packages.find]
where = ["."]

[tool.setuptools.package-data]
capellini = ["data/**/*", "r_scripts/*.R"]
